move com.teragrep.net_01.channel.context.frame; to com.teragrep.rlp_0… (#213)
* move com.teragrep.net_01.channel.context.frame; to com.teragrep.rlp_03.frame;

* apply spotless
