+++ src/test/java/com/teragrep/rlp_03/frame/clocks/EndOfTransferClockTest.java
@@ -43,7 +43,7 @@
  * Teragrep, the applicable Commercial License may apply to this file if you as
  * a licensee so wish it.
  */
-package com.teragrep.net_01.channel.context.frame.clocks;
+package com.teragrep.rlp_03.frame.clocks;
 
 import com.teragrep.rlp_03.frame.fragment.Fragment;
 import com.teragrep.rlp_03.frame.fragment.clocks.EndOfTransferClock;
